Allow using `../` as environment path
The main problem is that Tanka's standard is to put environments under the `environment/<env>` directory. This is mostly
fine, which is why this plugin implements it directly. But there are cases where an environment is just a bunch of files
in a repository, without any subdirectories. In that case we need to be able to look into the root directory, i.e.
name our environment `../`. This does not work directly with Tanka, so we need to resolve the path and give it to Tanka.